+++ CHANGES.md
@@ -6,6 +6,8 @@ Version 0.1.1
 
 To be released.
 
+ -   Fixed an unexpected error that lack of `-o`/`--output-dir` option (which
+     is required) had caused.
  -   Removed a debug-purpose `console.log()` from web test runner script
      (*runner.js*).